pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.11.0",
    "google-generativeai>=0.8.5",
    "numpy>=2.2.6",
    "openai>=1.82.1",
    "pdfplumber>=0.11.6",
    "pypdf2>=3.0.1",
    "streamlit>=1.45.1",
]
